requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
werkzeug==2.3.7
mysqlclient==2.2.0
python-dotenv==1.0.0
requests==2.31.0
openpyxl==3.1.2
xhtml2pdf==0.2.8
chart.js==0.1.0
groq==0.3.0